requirements.txt
asgiref==3.5.2
Django==4.0.1
django-environ==0.9.0
psycopg2-binary==2.9.5
djangorestframework==3.14.0
social-auth-app-django==5.0.0
sqlparse==0.4.3
tzdata==2022.7
django-filter~=22.1
marshmallow~=3.19.0
requests~=2.28.1
marshmallow-dataclass==8.5.11
pytest-django~=4.5.2
pytest~=7.2.1
Faker~=16.6.0
factory-boy~=3.2.1